# backend/openapi.yaml
openapi: 3.0.3
info:
  title: LAB Token Backend API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: Service health
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  chain:
                    type: string
                  totalSupply:
                    type: string
  /config:
    get:
      summary: Get token configuration and fees
      responses:
        "200":
          description: Token configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  symbol:
                    type: string
                  decimals:
                    type: integer
                  totalSupply:
                    type: string
                  wallets:
                    type: object
                    additionalProperties:
                      type: string
                  fees:
                    type: object
                    additionalProperties:
                      $ref: "#/components/schemas/FeeConfig"
  /balance/{address}:
    get:
      summary: Get balance of address
      parameters:
        - in: path
          name: address
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Balance details
          content:
            application/json:
              schema:
                type: object
                properties:
                  address:
                    type: string
                  balance:
                    type: string
                  formatted:
                    type: string
        "400":
          description: Invalid address
  /set-fees:
    post:
      summary: Update fee configuration
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                feeType:
                  $ref: "#/components/schemas/FeeType"
                config:
                  $ref: "#/components/schemas/FeeConfig"
      responses:
        "200":
          description: Transaction receipt
  /set-fee-wallets:
    post:
      summary: Update fee wallets
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                platformWallet:
                  type: string
                creatorWallet:
                  type: string
                royaltyWallet:
                  type: string
      responses:
        "200":
          description: Transaction receipt
  /slot-buy:
    post:
      summary: Execute slot buy transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipient:
                  type: string
                amount:
                  type: string
                referrer:
                  type: string
      responses:
        "200":
          description: Transaction receipt
  /direct-commission:
    post:
      summary: Execute direct commission transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seller:
                  type: string
                amount:
                  type: string
      responses:
        "200":
          description: Transaction receipt
  /royalty-transfer:
    post:
      summary: Execute royalty transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipient:
                  type: string
                amount:
                  type: string
      responses:
        "200":
          description: Transaction receipt
  /super-royalty-transfer:
    post:
      summary: Execute super royalty transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipient:
                  type: string
                amount:
                  type: string
                payees:
                  type: array
                  items:
                    type: string
                bps:
                  type: array
                  items:
                    type: integer
      responses:
        "200":
          description: Transaction receipt
  /creator-transfer:
    post:
      summary: Execute creator transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recipient:
                  type: string
                amount:
                  type: string
      responses:
        "200":
          description: Transaction receipt
  /flash-transfer:
    post:
      summary: Execute flash transfer
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                amount:
                  type: string
      responses:
        "200":
          description: Transaction receipt
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    FeeType:
      type: string
      enum:
        - slotBuy
        - directCommission
        - royaltyTransfer
        - superRoyaltyTransfer
        - creatorTransfer
        - flashTransfer
    FeeConfig:
      type: object
      properties:
        platformFeeBps:
          type: integer
        creatorFeeBps:
          type: integer
        royaltyFeeBps:
          type: integer
        referrerFeeBps:
          type: integer